feat: AI credit system, token tracking, review system, lab improvements, IPD enhancements
- AI credit system: purchase packs, deduct on AI use, token tracking per request
- Token usage analytics: input/output tokens, cost USD/INR per feature
- Credit balance shown in vet navbar with live updates
- Review system: auto-send on appointment completion, GMB integration
- Org review dashboard with ratings analytics
- Clinic edit functionality with GMB review URL
- Lab improvements: city-based matching, test catalog preview, clinic-level assignment
- Lab test standardization with master directory and standard codes
- Awaiting lab results appointment status
- IPD treatment auto-fetch from KB with dose calculation
- IPD history inline in pet history modal
- Drug submission workflow for org-submitted drugs
- Vet hiring platform, schedule management, vaccination records
- Case sheet UI improvements with clear section differentiation
- Switch to clinic panel for authorized vets
- Various UI fixes across org, clinic, and vet panels

+++ app/Models/Appointment.php
@@ -84,9 +84,9 @@ class Appointment extends Model
             return "{$baseYears}y {$baseMonths}m";
         }
 
-        $diffMonths = $recordedAt->diffInMonths($visitDate);
+        $diffMonths = (int) $recordedAt->diffInMonths($visitDate);
 
-        $totalMonths = ($baseYears * 12) + $baseMonths + $diffMonths;
+        $totalMonths = (int) (($baseYears * 12) + $baseMonths + $diffMonths);
 
         $years  = intdiv($totalMonths, 12);
         $months = $totalMonths % 12;
